Enhancements for Release 1.0
Addressing issues 23 and 22 from the MagRocket Backend Issues List...  Enhanced logging in development mode to the System Log to provide easier debugging for setup and testing of issues.  Ability to mark issues as FREE which will allow MagRocket to populate a free trial sample issue(s) in Baker while simultaneously allowing the use of Paid Subscriptions and Single Issue Purchases.