Issue Date Fix
Modified the Production level issue date criteria to allow for the start date to begin at the first day of the month that the subscription was purchased in and the end date to be the last day of the month that the subscription expires.  This is to cover the case where issues for the month were created prior to the user subscribing, but they should still have access to those issues.